refactor: テーマをModelとして切り出し
