+++ src/component/Theme/Theme.tsx
@@ -1,26 +1,7 @@
+import { themes } from "@models/Theme";
 import React from "react";
 
 import { SelectRow } from "./SelectRow";
-
-type TThemesTypes = {
-  type: string;
-  label: string;
-};
-
-const themes: TThemesTypes[] = [
-  {
-    type: "os",
-    label: "OSの設定に合わせる",
-  },
-  {
-    type: "light",
-    label: "ライト",
-  },
-  {
-    type: "dark",
-    label: "ダーク",
-  },
-];
 
 export const Theme: React.VFC = () => (
   <div className={"mt-8"}>
